+++ openbas-front/src/admin/components/workspaces/custom_dashboards/widgets/configuration/list/WidgetPerspectiveSelection.tsx
